requirements.txt
# Web Framework
fastapi>=0.68.0
uvicorn>=0.15.0

# Database
mysql-connector-python>=8.0.26
SQLAlchemy>=1.4.23
alembic>=1.7.1

# Cache
redis>=4.3.4

# Message Queue
pika>=1.2.0

# Utils
python-dotenv>=0.19.0
loguru>=0.5.3
pydantic>=1.8.2

# Video Processing
opencv-python>=4.5.3
numpy>=1.21.2

# Audio Processing
librosa>=0.8.1
soundfile>=0.10.3
torch>=1.9.0
torchaudio>=0.9.0
TTS